requirements.txt
phidata
streamlit
openai
python-dotenv
